+++ apiPenduraAi/src/main/java/com/ufc/apiPenduraAi/dtos/divida/CreateDividaDTO.java
@@ -1,4 +1,4 @@
 package com.ufc.apiPenduraAi.dtos.divida;
 
-public record CreateDividaDTO(String nome, String valor, Long user_id) {
+public record CreateDividaDTO(String nome, String valor, int user_id) {
 }
