+++ apiPenduraAi/src/main/java/com/ufc/apiPenduraAi/dtos/divida/ReturnDividasDTO.java
@@ -1,4 +1,4 @@
 package com.ufc.apiPenduraAi.dtos.divida;
 
-public record ReturnDividasDTO(Long id, String cliente, String valor) {
+public record ReturnDividasDTO(int id, String cliente, String valor) {
 }
